+++ TM4C123_GPIO_LCD1602/xDriver_MCU/Driver_Source/NVIC.c
@@ -46,8 +46,8 @@ inline NVIC_nENABLE NVIC__enGetEnableIRQ(NVIC_nSTIR enIRQ)
 
     if((uint8_t)enIRQ <=NVIC_IRQ_MAX)
     {
-        u32IsrBit=enIRQ%32;
-        u32IsrIndex=enIRQ/32;
+        u32IsrBit=(uint32_t)enIRQ%32u;
+        u32IsrIndex=(uint32_t)enIRQ/32u;
         if((NVICw->ISER[u32IsrIndex]&(1<<u32IsrBit)) == (1<<u32IsrBit))
         {
             enStatus= NVIC_enENABLE;
@@ -65,8 +65,8 @@ inline NVIC_nSTATUS NVIC__enSetEnableIRQ(NVIC_nSTIR enIRQ, NVIC_nPRIORITY enPrio
 
     if((uint8_t)enIRQ <=NVIC_IRQ_MAX)
     {
-        u32IsrBit=enIRQ%32;
-        u32IsrIndex=enIRQ/32;
+        u32IsrBit=(uint32_t)enIRQ%32u;
+        u32IsrIndex=(uint32_t)enIRQ/32u;
         enStatus = NVIC__enSetPriorityIRQ(enIRQ,enPriority);
         if(enStatus == NVIC_enOK)
         {
@@ -85,8 +85,8 @@ inline NVIC_nSTATUS NVIC__enClearEnableIRQ(NVIC_nSTIR enIRQ)
 
     if((uint8_t)enIRQ <=NVIC_IRQ_MAX)
     {
-        u32IsrBit=enIRQ%32;
-        u32IsrIndex=enIRQ/32;
+        u32IsrBit=(uint32_t)enIRQ%32u;
+        u32IsrIndex=(uint32_t)enIRQ/32u;
         NVICw->ICER[u32IsrIndex]|=(1<<u32IsrBit);
         enStatus = NVIC_enOK;
     }
@@ -102,8 +102,8 @@ inline NVIC_nPENDING NVIC__enGetPendingIRQ(NVIC_nSTIR enIRQ)
 
     if((uint8_t)enIRQ <=NVIC_IRQ_MAX)
     {
-        u32IsrBit=enIRQ%32;
-        u32IsrIndex=enIRQ/32;
+        u32IsrBit=(uint32_t)enIRQ%32u;
+        u32IsrIndex=(uint32_t)enIRQ/32u;
         if((NVICw->ISPR[u32IsrIndex]&(1<<u32IsrBit)) == (1<<u32IsrBit))
         {
             enStatus= NVIC_enPENDING;
@@ -122,8 +122,8 @@ inline NVIC_nSTATUS NVIC__enSetPendingIRQ(NVIC_nSTIR enIRQ)
 
     if((uint8_t)enIRQ <=NVIC_IRQ_MAX)
     {
-        u32IsrBit=enIRQ%32;
-        u32IsrIndex=enIRQ/32;
+        u32IsrBit=(uint32_t)enIRQ%32u;
+        u32IsrIndex=(uint32_t)enIRQ/32u;
         NVICw->ISPR[u32IsrIndex]|=(1<<u32IsrBit);
         enStatus= NVIC_enOK;
     }
@@ -140,8 +140,8 @@ inline NVIC_nSTATUS NVIC__enClearPendingIRQ(NVIC_nSTIR enIRQ)
 
     if((uint8_t)enIRQ <=NVIC_IRQ_MAX)
     {
-        u32IsrBit=enIRQ%32;
-        u32IsrIndex=enIRQ/32;
+        u32IsrBit=(uint32_t)enIRQ%32u;
+        u32IsrIndex=(uint32_t)enIRQ/32u;
         NVICw->ICPR[u32IsrIndex]|=(1<<u32IsrBit);
         enStatus= NVIC_enOK;
     }
@@ -158,8 +158,8 @@ inline NVIC_nACTIVE NVIC__enGetActiveIRQ(NVIC_nSTIR enIRQ)
 
     if((uint8_t)enIRQ <=NVIC_IRQ_MAX)
     {
-        u32IsrBit=enIRQ%32;
-        u32IsrIndex=enIRQ/32;
+        u32IsrBit=(uint32_t)enIRQ%32u;
+        u32IsrIndex=(uint32_t)enIRQ/32u;
         if((NVICw->IABR[u32IsrIndex]&(1<<u32IsrBit)) == (1<<u32IsrBit))
         {
             enStatus= NVIC_enACTIVE;
